requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
httpx>=0.28.1
numpy>=1.26.0
faiss-cpu>=1.8.0
